[AIRLFOW-YYY] Move dag commands to separate file
